chore: remove unused bitcoin code, cleanup deps, feature gate parser
