+++ crates/bitcoin/src/utils.rs
@@ -1,6 +1,8 @@
+use primitive_types::{H256, U256};
 use sha2::{Digest, Sha256};
-use sp_core::{H256, U256};
-use sp_std::{prelude::*, vec};
+
+#[cfg(not(feature = "std"))]
+use alloc::{vec, vec::Vec};
 
 use crate::types::H256Le;
