+++ parachain/runtime/runtime-tests/Cargo.toml
@@ -146,7 +146,7 @@ hex = '0.4.2'
 mocktopus = "0.8.0"
 serde_json = "1.0"
 
-bitcoin = { path = "../../../crates/bitcoin" }
+bitcoin = { path = "../../../crates/bitcoin", features = ["parser"] }
 
 [features]
 default = ["std"]
